requirements.txt
langchain
chromadb
pymupdf
whisper
pytube
sentence-transformers
torch
pillow
tqdm
openai
transformers
peft
bitsandbytes
scikit-learn
llama-index
crewai